Move allow wallet notification to lib.
